+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.3.0
 
- - Added metrics to the qualification methodology
+ - Added metrics (including tests) to the qualification methodology
    - Attribute completeness
    - Non-default attribute accuracy
    - Object type coverage
@@ -8,3 +8,5 @@
    - Temporal completeness
    - Class completeness
    - Duplicate record rate
+- Refactored metrics to be more modular and reusable
+- Added tests for existing metrics, metric-class and metrics-manager